Web App/requirements.txt
django
djangorestframework
mysqlclient
requests
urllib3